Tripod Gun and Turret Gun now looks at JSON def for their rotation limits. if there is no value in JSON(that equals 0), they use default values. If the limit needs to be 0, JSON must have some value close enough to it. Also Changed limit variables to float instead of int since methods related to them return floats.

+++ src/main/java/minecrafttransportsimulator/jsondefs/JSONPart.java
@@ -88,10 +88,10 @@ public class JSONPart extends AJSONMultiModel<JSONPart.PartGeneral>{
     	public int fireDelay;
     	public int reloadTime;
     	public int muzzleVelocity;
-    	public int minPitch;
-    	public int maxPitch;
-    	public int minYaw;
-    	public int maxYaw;
+    	public float minPitch;
+    	public float maxPitch;
+    	public float minYaw;
+    	public float maxYaw;
     	public float diameter;
     	public float length;
     }
